made several changes
added another way for the player to shoot by being able to use either spacebar or pressing/holding down the left mouse button to fire.  The font was changed so it wasn't using the default font, and I removed the text that would display what the controls were since the description for my itch.io build will show what controls are being used in the description.